apps/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
celery>=5.3.6
redis>=5.0.1
yt-dlp>=2023.11.16
ffmpeg-python>=0.2.0
Pillow>=10.4.0
httpx>=0.26.0
python-multipart>=0.0.9
pytesseract>=0.3.10
